Fullstack app specification for quiz app

+++ React.NET/Models/UserScoreDto.cs
@@ -8,6 +8,6 @@ namespace React.NET.Models
     public class UserScoreDto
     {
         public string Username { get; set; }
-        public int Score { get; set; }
+        public Decimal Score { get; set; }
     }
 }
